Just straight up code
If you want to copy what's in here and add it to your own FarmhouseA.cs files. I will likely be making many more for all of the other production houses, this was mostly just to test everything, but it has a lot of work done already. I'll be adding straight code like this as well as updating the ZIP once we start doing asset integration.